# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Advisor Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "jstojiljkovic941@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/flight-advisor/v1/airports/import": {
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Import airports into storage",
                "description": "",
                "operationId": "import",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "airports"
                                ],
                                "properties": {
                                    "airports": {
                                        "description": "Airports to import",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Returns when validation fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Imports airports inside a database",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Imports airports inside a database",
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}